Move isSharedEncoding to TritonGPUIR.
This breaks a cyclic dependency between TritonAnalysis and TritonGPUIR
(see #1649).
